scanner: stream scanner module
The text/event-stream format includes mixed line endings, and is not easily
consumed without a custom scanner.  This scanner tokenizes the stream as well;
working under the assumption that it needs to be able to detect protocol
errors.  However, according to the spec, there is literally no response which
is not conformant to the protocol; however later this approach will allow
prompt detection of streams which are *probably* garbage.

I also just wanted to get more familiar with the scanner API; so there is that.
I might replace this with a simple equivalent to bufio.ScanLines that allows
lines terminated by just a '\r' (which is ambiguous and requires scanner
state).