sink: a basic SSE handler func for net/http servers
In order to write SSE handlers, you need to have something
to call from your endpoint or router.  This API allows
handlers which are capable of delivering events which know
how to marshal to byte form via a GetData method.  Using an
interface type allows some flexibility for using interface
detection as a form of side-channel.

As events will be Sink events, change the ‘LastEventID’ field
to just ‘EventID’