sink: implement keepalives
Keepalives can be useful to know when a device which is thought
to be connected has in reality disappeared ahead of the time
you next need to send them a message.  They have a downside in
that transient network errors may result in breakages you didn’t
want.

Add a default 1 minute keepalive, and two ways to specify it:
a method or an event side channel.