client: emit open/close events
These events are useful to know when a connection bounces; for
instance, if the endpoint does not support at least once delivery
of messages, then it will be necessary to refresh the current
state of the resource from the server on open.  And closes can be
used to notify the user that the connection is down.