protocol.go: make a copy of the token
If a large event is read (more than about 4k), then the byte
array returned by Scan is recycled, and so the slice in the
'field' variable gets replaced by the first few bytes of the
second scanned chunk.  This means it doesn't match 'data'
any more, and so the event is swallowed in the client.  Make
a real copy of the token before continuing.